+++ app/(authentication)/components/consts.ts
@@ -1,4 +1,5 @@
 export const TEXTS = {
     signedInAs: "Signed in as",
     signOut: "Sign me out",
+    usernamePrefix: "Welcome,",
 }
